refactor function storeRegisterInfo

+++ server/app/Http/Controllers/Api/InternshipGraduationController.php
@@ -155,12 +155,14 @@ class InternshipGraduationController extends Controller
     public function storeRegisterInfo(Request $request)
     {
         $internship_graduation_id = $request->input('internship_graduation_id');
-        $companiesDB = RegisterIntershipCompany::where('internship_graduation_id', $internship_graduation_id)
-            ->leftJoin('company_position_detail', 'company_position_detail.register_internship_company_id', '=', 'register_internship_company.register_internship_company_id');
         $companiesUpdate = collect($request->input('companies'));
+        
+        $existingCompanies = RegisterIntershipCompany::where('internship_graduation_id', $internship_graduation_id)
+            ->leftJoin('company_position_detail', 'register_internship_company.register_internship_company_id', '=', 'company_position_detail.register_internship_company_id');
 
+        // Tách thành mảng company_id để xác định những phần tử nào cần thêm / cần xoá / cần chỉnh sửa
         $listIdUpdate = $companiesUpdate->pluck('company_id');
-        $listIdCompany = $companiesDB->pluck('company_id');
+        $listIdCompany = $existingCompanies->pluck('company_id');
 
         // tách dữ liệu
         $companiesCreate = $listIdUpdate->diff($listIdCompany)->unique();
@@ -169,22 +171,25 @@ class InternshipGraduationController extends Controller
 
         // intersect
         foreach ($intersect as $company) {
+            // Xử lý dữ liệu - Tách ra danh sách vị trí cần thêm / cần xoá
             $positionIdUpdate = [];
             $positionList = $companiesUpdate->where('company_id', $company)->first()['positions'];
             foreach ($positionList as $position) {
                 array_push($positionIdUpdate, $position['position_id']);
             }
             $positionIdDB = [];
-            foreach ($companiesDB->get() as $position) {
+            foreach ($existingCompanies->get() as $position) {
                 if ($position['company_id'] === $company) {
                     array_push($positionIdDB, $position['position_id']);
                 }
             };
 
+            // Lọc danh sách id vị trí
             $positionCreate = array_diff($positionIdUpdate, $positionIdDB);
             $positionDelete = array_diff($positionIdDB, $positionIdUpdate);
             $positionIntersect = array_intersect($positionIdUpdate, $positionIdDB);
 
+            // Cập nhật trường is_Interview
             $company_record_update = $companiesUpdate->where('company_id', $company)->first();
             $isInternview = $company_record_update['company_isInterview'];
 
@@ -194,17 +199,17 @@ class InternshipGraduationController extends Controller
                 'company_isInterview' => $isInternview
             ]);
 
-            // 
+            // Cập nhật vị trí
             $arrPo = $company_record_update['positions'];
             foreach ($positionIntersect as $position_id) {
                 $position_quantity = 0;
-                foreach ($arrPo as $po) {
-                    if ($po['position_id'] == $position_id) {
-                        $position_quantity = $po['position_quantity'];
+                foreach ($arrPo as $positionItem) {
+                    if ($positionItem['position_id'] == $position_id) {
+                        $position_quantity = $positionItem['position_quantity'];
                         break;
                     }
                 }
-                $query = clone $companiesDB;
+                $query = clone $existingCompanies;
                 $instance = $query->where('position_id', $position_id)->first();
                 $com = CompanyPositionDetail::find($instance->company_position_detail_id);
                 $com->update([
@@ -212,25 +217,25 @@ class InternshipGraduationController extends Controller
                 ]);
             }
 
-            // Delte
+            // Delete position
             foreach ($positionDelete as $position_id) {
-                $query = clone $companiesDB;
+                $query = clone $existingCompanies;
                 $instance = $query->where('position_id', $position_id)->first();
                 $com = CompanyPositionDetail::find($instance->company_position_detail_id);
-                $com->delete();
+                if ($com) $com->delete();
             }
 
             // Create
 
             foreach ($positionCreate as $position_id) {
-                foreach ($arrPo as $po) {
-                    if ($po['position_id'] == $position_id) {
-                        $query = clone $companiesDB;
+                foreach ($arrPo as $positionItem) {
+                    if ($positionItem['position_id'] == $position_id) {
+                        $query = clone $existingCompanies;
                         $instance = $query->where('company_id', $company)->first();
-                        $com->create([
+                        CompanyPositionDetail::create([
                             "register_internship_company_id" => $instance->register_internship_company_id,
-                            "position_id" => $po['position_id'],
-                            "position_quantity" => $po['position_quantity'],
+                            "position_id" => $positionItem['position_id'],
+                            "position_quantity" => $positionItem['position_quantity'],
                         ]);
                         break;
                     }
@@ -240,7 +245,7 @@ class InternshipGraduationController extends Controller
 
         // Delete company
         foreach ($companiesDelete as $company) {
-            $instance = $companiesDB->where('company_id', $company);
+            $instance = $existingCompanies->where('company_id', $company);
             $instance->delete();
         }
